requirements.txt
requests
psycopg2-binary
python-dotenv
prettytable
streamlit
pandas
